requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
openai>=1.30.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
pyyaml>=6.0.1
python-dotenv>=1.0.1
pydantic>=2.7.0
numpy>=1.26.0
